+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [unreleased]
+## [1.9.0] - 2018-11-29
 - **Added** `propsComputed` api for props manipulation
 - **Fixed** `parentNode` is `null` when using scoped style without others tags